lesson26/requirements.txt
requests==2.25.1
pytest==6.2.2
allure-pytest==2.8.36
selenium==3.141.0
pytest-xdist==2.2.1
